extract combining utils into library
